modify delete friend

+++ FriendRelation/views.py
@@ -40,13 +40,14 @@ def delete_friend(req: HttpRequest):
                 name_list = [username, friend_name]
                 friend = Friend.objects.filter(user_name=name_list[i], friend_name=name_list[1-i]).first()
 
-                flist = FriendList.objects.get(user_name=name_list[i])
-                for name in flist.friend_list:
-                    if friend.name_list[1-i] == name:
-                        flist.friend_list.remove(name)
-                        break
-                flist.save()
-                friend.delete()
+                if not friend is None:
+                    flist = FriendList.objects.get(user_name=name_list[i])
+                    for name in flist.friend_list:
+                        if friend.name_list[1-i] == name:
+                            flist.friend_list.remove(name)
+                            break
+                    flist.save()
+                    friend.delete()
 
             return JsonResponse({
                 'code': 0,
